fix line split
